deployment/requirements.txt
streamlit
pandas 
numpy
seaborn 
matplotlib
plotly 
numpy 
Pillow 
scikit.learn
tensorflow
keras
opencv-python